requirements.txt
alembic==1.16.5
asgiref==3.8.1
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.2.1
distlib==0.3.8
Django==5.1.3
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.14.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
platformdirs==4.2.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dotenv==1.1.1
pytube==15.0.0
pytz==2023.3.post1
requests==2.32.3
SQLAlchemy==2.0.43
sqlparse==0.5.2
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.4.0
virtualenv==20.26.2
Werkzeug==3.1.3
yt-dlp==2024.12.23
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1

